use same style for labels
